requirements.txt
anyio==3.6.2
astroid==2.15.2
autopep8==2.0.2
click==8.1.3
dill==0.3.6
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
pylint==2.17.2
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a33
sqlmodel==0.0.8
starlette==0.26.1
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
uvicorn==0.21.1
wrapt==1.15.0
